fix: Rewrite to interpret multiaddr from right to left (#120)
I wanted to support `/tls/ws`, `/tls/http`, and `/tls/sni/example.com/http`. Unfortunately the existing code interprets these multiaddrs left to right in a reduce loop. These multiaddrs (and multiaddrs in general) need to be parsed right to left. This refactors the code to so that it is interpreted right to left.

It's a bit verbose, but straightforward.

This is somewhat blocked on https://github.com/multiformats/js-multiaddr/pull/312 (Not strictly blocked, but would be good to bump the dep when that is merged).

This is needed because js-libp2p-websockets uses this library, and without this we can't dial addresses that use `/tls/ws` or `/tls/sni/example.com/ws...`

---------

Co-authored-by: achingbrain <alex@achingbrain.net>